+++ packages/local-webrtc-browser-sdk/src/sdk/host-battle-sdk.ts
@@ -10,11 +10,12 @@ import {
   startGBraverBurst,
 } from "gbraver-burst-core";
 import { nanoid } from "nanoid";
-import { EMPTY, Observable } from "rxjs";
+import { from, mergeMap, Observable, Subject, take, takeUntil } from "rxjs";
 
 import { SendCommand } from "../webrtc/guest/guest-message";
 import { sendHostMessage } from "../webrtc/host/host-message";
 import { receiveSendCommand } from "../webrtc/host/receive-send-command";
+import { notifyConnectionFailed } from "../webrtc/notify-connection-failed";
 import { BattleSDK } from "./battle-sdk";
 import { HostWebRTCConnectionManager } from "./host-webrtc-connection-manager";
 
@@ -34,6 +35,8 @@ export class HostBattleSDK implements BattleSDK {
   #webRTCConnection: HostWebRTCConnectionManager;
   /** ゲストからのコマンド受信プロミス */
   #sendCommandPromise: Promise<SendCommand>;
+  /** dataChannel.sendの例外通知ストリーム */
+  #sendExceptionSubject: Subject<unknown> = new Subject<unknown>();
 
   /**
    * コンストラクタ
@@ -93,11 +96,16 @@ export class HostBattleSDK implements BattleSDK {
     const dataChannel =
       await this.#webRTCConnection.getOrCreateConnection().dataChannelPromise;
     this.#sendCommandPromise = receiveSendCommand(dataChannel, this.flowID);
-    sendHostMessage(dataChannel, {
-      type: "battle-progressed",
-      flowID: this.flowID,
-      update,
-    });
+    try {
+      sendHostMessage(dataChannel, {
+        type: "battle-progressed",
+        flowID: this.flowID,
+        update,
+      });
+    } catch (error) {
+      this.#sendExceptionSubject.next(error);
+      throw error;
+    }
     return update;
   }
 
@@ -106,7 +114,12 @@ export class HostBattleSDK implements BattleSDK {
    * @returns 通知ストリーム
    */
   suddenlyBattleEndNotifier(): Observable<unknown> {
-    // TODO 中身を実装する
-    return EMPTY;
+    return from(
+      this.#webRTCConnection.getOrCreateConnection().connectionPromise,
+    ).pipe(
+      mergeMap((connection) => notifyConnectionFailed(connection)),
+      takeUntil(this.#sendExceptionSubject),
+      take(1),
+    );
   }
 }